Работаю с вкладкой Дети Работников

+++ src/main/java/com/quasar/hibernateh2/dao/entity/WorkersChild.java
@@ -14,24 +14,24 @@ import javax.persistence.Table;
 public class WorkersChild extends Model implements Serializable {
     
     @Column(name = "id_worker")
-    private int id_worker;
+    private Long id_worker;
     
     @Column(name = "id_child")
-    private int id_child;
+    private Long id_child;
 
-    public int getId_worker() {
+    public Long getId_worker() {
         return id_worker;
     }
 
-    public void setId_worker(int id_worker) {
+    public void setId_worker(Long id_worker) {
         this.id_worker = id_worker;
     }
 
-    public int getId_child() {
+    public Long getId_child() {
         return id_child;
     }
 
-    public void setId_child(int id_child) {
+    public void setId_child(Long id_child) {
         this.id_child = id_child;
     }
